backend/user_service/requirements.txt
Flask==2.3.2  
mysql-connector-python==8.1.0 
PyJWT
datetime
cryptography